Optimize the BFS strategy for 2016 day 11 by disregarding duplicate states
eg      1
      /   \
     2     3
    / \   / \
   3   4 5   2

In this scenario we record we've seen the 1, the 2, then the 3.
The second 3 (on the third level) is viable, but not optimal (it's shorter to get to 3 by going 1-3 than 1-2-3)
The second 2 (on the third level) is also not optimal (1-2 is better than 1-3-2).

So the number of nodes moving to the third level has halved, from 4 to 2.

When running it against the sample set in the test now:

             Old:          New:
             ----          ----
Level 1      1             1
Level 2      3             3
Level 3      11            11
Level 4      15            14
Level 5      32            24
Level 6      88            45
Level 7      203           77
Level 8      401           60
Level 9      820           50
Level 10     1627          45