2019 day 14, part 2
Initial optimization was to look for loops in the extras being fed
from one iteration to the next, which worked for the sample input 3
and 4 - but not 5 and not the real input, there were no loops.

Eventually I worked out that there must be at least (ore / ore_per_fuel)
fuel available, and worked from there.