Rewrite 2016 day 10
It didn't work in benchmarking, due to its use of a lot of GenServers and
a global registry for identifying them (couldn't be run in parallel).

Looking over the code I'm not sure why I even did that... so I rewrote it
to store state in a map that got passed around, and now it runs fine!